\docType{class}
\name{Avg.Fit-class}
\alias{Avg.Fit-class}
\title{A regression combination class}
\description{
Object of subclass \code{Avg.Fit} as created by the
\code{comboreg} functions
}
\details{
An object of the class `regcombo' has the following slots:
\itemize{ \item \code{avg.fit} The mean R-squared value in
all combos of coefficients }
}
\author{
Emily Moore: \email{emily.moore@wustl.edu}
}

