\docType{data}
\name{exampleDataset}
\alias{exampleDataset}
\title{Example Data}
\description{
Four exp,anatory variables and one dependent variable
}
\details{
The variables included in the dataset are: \itemize{
\item\code{Variable1} A vector of random numbers
\item\code{Variable2} Another vector of random numbers
\item\code{Variable3} Another vector of random numbers
\item\code{Variable4} Another vector of random numbers
\item\code{Dependent Variable} Dependent variable made up
of combo of some of explanatory variables plus error }
}

