\docType{class}
\name{regcombo-class}
\alias{regcombo-class}
\title{A regression combination class}
\description{
Object of class \code{regcombo} as created by the
\code{comboreg} functions
}
\details{
An object of the class `regcombo' has the following slots:
\itemize{ \item \code{combo.coef} A list of coefficients
for each possible regression \item \code{combo.fit} A
vector of R-Squared values \item \code{x} A matrix input of
covariates \item \code{y} The dependent variable input }
}
\author{
Emily Moore: \email{emily.moore@wustl.edu}
}

