\docType{package}
\name{RegComboPack}
\alias{RegComboPack}
\alias{RegComboPack-package}
\title{RegComboPack}
\description{
The RegComboPack package performs regressions on all combos
of covariates
}
\examples{
\dontrun{
demo(comboreg)
demo(comboreg.tval)
}
}
\author{
Emily Moore: \email{emily.moore@wustl.edu}
}

